Server: Test not just PONAPI::Client, but also HTTP::Tiny and cURL
...well, partially, anyway.  We still use PONAPI::Client for almost
everything, but the talking to the server now is now tested with
Hijk, HTTP::Tiny, and cURL (command-line version, not WWW::cURL).

cURL tests are 'special' -- since they are the most likely to become
as false-positives, they won't fail the test suite, but instead, will
get auto-skipped.  We can decide in a stable release if we want these
to be author tests, enabled everywhere, or removed.