Removed Scalar::Util as a direct dependency
Mostly busywork, really. We didn't need to import blessed() in the
classes, since Moose provided that, and we were over-testing by
using blessed() -- we don't really care if the DAO returns a
plain or blessed hashref, as long as it can be encoded to JSON.

Finally, the two uses of looks_like_number() in MockDB were buggy
(they didn't check the *private* return value of the function
to ensure that we had an integer, meaning page=1.5 was a valid entry)
so just replacing them with a simple regex is the way to go.